Update to prevent false positives
Before, Get-writable only checked the permissions of the directory and not the files themselves. This would result in false positives for directories you had permissions over but that contained files you could not right. It could have also caused false negatives if there were directories that had no write/modify permissions but the files themselves did.

Because this now checks each individual files, it takes a bit longer to run than before